requirements.txt
# requirements.txt

Flask
Flask-Cors
numpy
opencv-python
Pillow
requests
gunicorn
PyMuPDF
tqdm
rich

imgkit
beautifulsoup4
google-genai
python-dotenv
